Send uploaded dictionary (json) to the server
