+++ ui/src/xmlEditor/hur/sendToTheServer.ts
@@ -1,0 +1,9 @@
+import { updateHurrianDictionaryUrl } from '../../urls';
+
+export function sendMorphologicalAnalysisToTheServer(word: string, analysis: string) {
+  const formData = new FormData();
+  formData.append('word', word);
+  formData.append('analysis', analysis);
+
+  fetch(updateHurrianDictionaryUrl, {method: 'POST', body: formData});
+}
